use smallint for rating score, really

+++ migrations/406-signed-rating-scores.sql
@@ -1,1 +1,1 @@
-ALTER TABLE ratings CHANGE COLUMN score score smallint;
+ALTER TABLE ratings CHANGE COLUMN score score tinyint(1);
